refactor: ai/figma 패키지 내 외부 API 클라이언트를 client/ 서브패키지로 분리 (#14)
- ai/ClaudeApiClient, ClaudeApiProperties → ai/client/
- figma/FigmaApiClient, FigmaApiProperties, FigmaNode, FigmaNodeResponse → figma/client/
- figma/ 루트에는 내부 도메인 모델(FigmaDesignContext, FigmaDesignExtractor 등)만 잔류
- FigmaDesignExtractor, ReactGenerateService import 경로 업데이트
