+++ admin/src/main/java/com/example/admin_demo/domain/reactgenerate/ai/client/ClaudeApiClient.java
@@ -1,4 +1,4 @@
-package com.example.admin_demo.domain.reactgenerate.ai;
+package com.example.admin_demo.domain.reactgenerate.ai.client;
 
 import com.example.admin_demo.global.exception.InternalException;
 import java.time.Duration;
