+++ admin/src/main/java/com/example/admin_demo/domain/reactgenerate/figma/client/FigmaNode.java
@@ -1,4 +1,4 @@
-package com.example.admin_demo.domain.reactgenerate.figma;
+package com.example.admin_demo.domain.reactgenerate.figma.client;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
