+++ admin/src/main/resources/templates/pages/react-generate/react-generate-script.html
@@ -119,9 +119,6 @@
 
     // ─── 승인 요청 ───────────────────────────────────────────
     $('#btnRequestApproval').on('click', function () {
-        alert('구현 예정입니다.');
-        return;
-
         if (!currentId) return;
         if (!confirm('승인을 요청하시겠습니까?')) return;
